Add 1.4 large Cert.

+++ spdmlib/src/requester/encap_certificate.rs
@@ -96,8 +96,8 @@ impl RequesterContext {
             .unwrap();
 
         let mut length = get_certificate.length;
-        if length > MAX_SPDM_CERT_PORTION_LEN as u16 {
-            length = MAX_SPDM_CERT_PORTION_LEN as u16;
+        if length > MAX_SPDM_CERT_PORTION_LEN as u32 {
+            length = MAX_SPDM_CERT_PORTION_LEN as u32;
         }
 
         let offset = get_certificate.offset;
